+++ qiskit/pulse/instructions/__init__.py
@@ -22,3 +22,4 @@ An instruction can be added to a :py:class:`~qiskit.pulse.Schedule`, which is a
 sequence of scheduled Pulse ``Instruction`` s over many channels.
 """
 from .instruction import Instruction
+from .play import Play
