+++ qiskit/pulse/instructions/play.py
@@ -1,0 +1,53 @@
+# -*- coding: utf-8 -*-
+
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2020.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""TODO"""
+from typing import Optional
+
+from qiskit.pulse.channels import PulseChannel
+
+from .instruction import Instruction
+
+
+class Play(Instruction):
+    """TODO"""
+
+    def __init__(self, pulse: Pulse, channel: PulseChannel, name: Optional[str] = None):
+        """TODO"""
+        self._pulse
+        self._channel = channel
+        super().__init__(pulse.duration, channel)
+
+    @property
+    def channel:
+        """TODO"""
+        return channel
+
+    @property
+    def pulse(self) -> Pulse:
+        """TODO"""
+        return pulse
+
+    @property
+    def pulse_name(self) -> str:
+        return pulse.name
+
+    @property
+    def operands(self):
+        """TODO"""
+        return # TODO
+
+    def __repr__(self):
+        """TODO"""
+        pass
